moved and created widgets
