--- src/Business/Browse.java
@@ -1,27 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package Business;
-
-import widgets.Widget;
-
-/**
- *
- * @author andt
- */
-//Browse is a collection of ProductContainers in different sizes.
-public class Browse extends Widget {
-
-//    private int height;
-//    private int width;
-    //on creation of a browser widget, size is to be defined first
-
-    public Browse(int width, int height) {
-        super.width = width;
-        super.height = height;
-    }
-
-
-}
